refactor(mocks): move mock error logs generator function to mocks from logger test file in reducers
re #7
