+++ libs/mocks/src/lib/mocks.ts
@@ -2,3 +2,4 @@ export * from './player';
 export * from './tournament';
 export * from './general';
 export * from './race';
+export * from './error-logs';
